Proposed Correction to primitive.js 201801270033
The function cylinder(...) is drawing incorrectly in version 0.6.0. Instead of the cylinder drawing a cylinder with a specified circle cap radius, and a   specified height, the function uses the specified radius and height arguments/parameters to draw an ellipse cap figure with a major radii, and a minor radii, and the height as specified radius. 
Prognosis: the line this._renderer.drawBuffersScaled is passing the specified parameters out of order. I have proposed a possible corrected order of these parameters.
This behavior of this function, drawing a cylinder with a circle cap with the specified radius, and the  specified height, is correct in the cloudfire version 0.5.9